Consolidate type exports, use UserRolesShape
